zayava/requirements.txt
aiogram==3.17.0
pillow==10.1.0
python-dotenv==1.0.0 